Add Stage 3 Weakness Audit Script for chart extraction evaluation
- Implemented a comprehensive script to audit the extraction pipeline on various chart types.
- The script generates detailed JSON and Markdown reports, including success rates, failure categories, and confidence distributions.
- Introduced data structures for storing audit results and summary statistics.
- Added functions for sampling charts, auditing individual charts, computing summaries, and generating reports.
- Integrated logging for better traceability during the audit process.

+++ src/core_engine/stages/s3_extraction/skeletonizer.py
@@ -654,11 +654,12 @@ class Skeletonizer:
         Returns:
             Stroke width map (2 * distance_map at skeleton pixels)
         """
-        stroke_width = np.zeros_like(distance_map)
+        stroke_width = np.zeros_like(distance_map, dtype=np.float64)
         
         # Stroke width = 2 * distance at skeleton
+        # [FIX] Cast to float64 to prevent overflow when distance_map is uint8
         mask = skeleton > 0
-        stroke_width[mask] = 2 * distance_map[mask]
+        stroke_width[mask] = 2.0 * distance_map[mask].astype(np.float64)
         
         return stroke_width
